Removed the refresh button.
It was too easy to click by accident. There might be a better place to put it (and there  should be an easy way for a user to start over with a blank page, and alg.garron.us alone might not autocomplete in the URL bar), but I'm taking it out for now.